python/pyproject.toml
[project]
name = "python"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "grpcio==1.59.0 ; python_full_version >= '3.7'",
    "grpcio-tools==1.59.0 ; python_full_version >= '3.7'",
    "kaikosdk==1.33.1",
    "protobuf==4.25.8 ; python_full_version >= '3.8'",
    "retrying==1.3.3",
    "setuptools==80.9.0 ; python_full_version >= '3.9'",
    "six==1.17.0 ; python_full_version >= '2.7'",
]
